finish refactor, remove empty actionhandlers from timer-button
