+++ src/styles/variables.scss
@@ -1,6 +1,7 @@
 // General Styles
 $bg-primary: #FDFDFD; // Primary Background Color
-$border-primary: 1px solid #D6D6D6; // Primary Border config
+$border-color: #D6D6D6;
+$border-primary: 1px solid $border-color; // Primary Border config
 $color-secondary: rgba(160, 175, 196, 1);
 $container-padding: .5rem;
 $item-padding: 18px 38px; // Standard Padding
