--- src/x-esg-timer-button/actionHandlers.js
@@ -1,7 +1,0 @@
-import { actionTypes } from '@servicenow/ui-core';
-
-const {COMPONENT_BOOTSTRAPPED} = actionTypes;
-
-export default {
-
-}
