+++ src/x-esg-timer-button/index.js
@@ -1,7 +1,6 @@
 import {createCustomElement} from '@servicenow/ui-core';
 import {snabbdom} from '@servicenow/ui-renderer-snabbdom';
 import {view} from './view';
-import actionHandlers from './actionHandlers';
 import styles from './styles.scss';
 
 createCustomElement('x-esg-timer-button', {
@@ -21,5 +20,4 @@ createCustomElement('x-esg-timer-button', {
 		projectData: {default: {}},
 		loadFonts: {default: true},
 	},
-	actionHandlers,
 });
