+++ src/x-esg-timer-container/components/AddProject.js
@@ -1,5 +1,3 @@
-
-
 export const AddProject = ({
                                 addProjectStatus, 
                                 projects,
