+++ src/x-esg-timer-container/components/Timestamp.js
@@ -1,5 +1,5 @@
-import {format, formatDistanceToNow, min} from 'date-fns';
-import { msToString, hhmmToSnTime, getUTCTime, toSnTime } from '../../x-esg-timer-button/helpers';
+import { hhmmToSnTime, getUTCTime, toSnTime } from '../../x-esg-timer-button/helpers';
+import { format } from 'date-fns';
 
 export const Timestamp = ({ 
                             stamp,
