+++ src/x-esg-timer-container/styles/styles.scss
@@ -1,13 +1,14 @@
 @import '@servicenow/sass-kit/host';
 @import '../../styles/variables.scss';
 
-// font classes
+// font and symbol classes
 @import './materialSymbols.scss';
 
 // Component/subcomponent modules;
 @import './outerButtons.scss';
 @import './todayContainer.scss';
 @import './projectItem.scss';
+@import './todayHeader.scss';
