+++ src/x-esg-timer-container/styles/todayHeader.scss
@@ -1,0 +1,43 @@
+@import '../../styles/variables.scss';
+
+.today-header {
+    padding: $item-padding;
+    font-size: $font-size-large;
+    width: 100%;
+    display: inline-grid;
+    grid-template-columns: 1fr 1fr 1fr;
+
+    .title {
+        font-weight: $font-weight-bold;
+        margin-right: 50px;
+    }
+
+    .header-date {
+        display: inline-grid;
+        width: 300px;
+        grid-template-columns: 1fr 6fr 1fr;
+        text-align: center;
+
+        .date-chevron {
+            transition: .4s all;
+            &:hover {
+                cursor: pointer;
+                filter: drop-shadow(0 0 .1rem rgb(95, 95, 95));
+            }
+
+            &.disabled {
+                color: $border-color;
+                cursor: default;
+
+                &:hover {
+                    filter: none;
+                    cursor: default;
+                }
+            }
+        }
+    }  
+
+    .today-total {
+        text-align: right;
+    }
+}
